backend/pyproject.toml
[tool.poetry]
name = "marketing-campaign-tracker"
version = "0.1.0"
description = "Marketing campaign monitoring application for Google Ads, Meta Ads, and Reddit Ads"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
google-ads = "^24.1.0"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
pandas = "^2.2.0"
cryptography = "^42.0.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
httpx = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
